I just realised the RTC isn't actually fixed. If you drop the supply voltage below threshold without letting it actually hit 0v the RTC is fine but if it does hit 0v the RTC resets. This doesn't happen in the Sparkfun v2.7 code. The Sparkfun code just works.
I think I have narrowed the issue down to the "setupSync()" function called in the setup() function of the OLA. When this function is commented out the RTC works (even when the voltage is dropped to 0v) but when you bring the function back in the RTC only works partially.

Its likely something to do with the interrupts we use for the logger and coordinator time keeping.

This is the last stable commit before I start messing around and trying to fix this problem.